+++ Docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Markdown ToolBox Change Log 📋
 
-## v1.2 *(current)* 🆕
+## v1.3 *(current)* 🆕
+#### 🚀 New Features:
+    • Added a button to enable dark mode.
+    • Added a button to reload the current webpage.
+
+## v1.2 🔄
 #### 🛠️ Fixes:
     • ChatGPT website did not allow login.